docker_mqtt/backend/preprocessing/requirements.txt
nltk==3.4.5
pandas==1.0.1
scikit_learn==0.23.2
xlrd==1.2.0
XlsxWriter==1.3.7
paho-mqtt